Move association-related code from Step to JwstStep
