+++ jwst/datamodels/container.py
@@ -8,10 +8,6 @@ from asdf import AsdfFile
 from astropy.io import fits
 from stdatamodels import DataModel, properties
 
-from ..associations import (
-    AssociationNotValidError,
-    load_asn)
-
 from .model_base import JwstDataModel
 from .util import open as datamodel_open
 from .util import is_association
@@ -180,6 +176,8 @@ class ModelContainer(JwstDataModel):
         filepath : str
             The path to an association file.
         """
+        # Prevent circular import:
+        from ..associations import AssociationNotValidError, load_asn
 
         filepath = op.abspath(op.expanduser(op.expandvars(filepath)))
         try:
