+++ jwst/stpipe/step.py
@@ -31,10 +31,9 @@ from . import config_parser
 from . import crds_client
 from . import log
 from . import utilities
+from .format_template import FormatTemplate
+
 from .. import __version_commit__, __version__
-from ..associations.load_as_asn import (LoadAsAssociation, LoadAsLevel2Asn)
-from ..associations.lib.format_template import FormatTemplate
-from ..associations.lib.update_path import update_key_value
 from ..datamodels import (ModelContainer, StepParsModel)
 from ..lib.class_property import ClassInstanceMethod
 from ..lib.suffix import remove_suffix
@@ -1115,45 +1114,6 @@ class Step(abc.ABC):
 
         return full_path
 
-    def load_as_level2_asn(self, obj):
-        """Load object as an association
-
-        Loads the specified object into a Level2 association.
-        If necessary, prepend `Step.input_dir` to all members.
-
-        Parameters
-        ----------
-        obj : object
-            Object to load as a Level2 association
-
-        Returns
-        -------
-        association : jwst.associations.lib.rules_level2_base.DMSLevel2bBase
-            Association
-        """
-        asn = LoadAsLevel2Asn.load(obj, basename=self.output_file)
-        update_key_value(asn, 'expname', (), mod_func=self.make_input_path)
-        return asn
-
-    def load_as_level3_asn(self, obj):
-        """Load object as an association
-
-        Loads the specified object into a Level3 association.
-        If necessary, prepend `Step.input_dir` to all members.
-
-        Parameters
-        ----------
-        obj : object
-            Object to load as a Level3 association
-
-        Returns
-        -------
-        association : jwst.associations.lib.rules_level3_base.DMS_Level3_Base
-            Association
-        """
-        asn = LoadAsAssociation.load(obj)
-        update_key_value(asn, 'expname', (), mod_func=self.make_input_path)
-        return asn
 
     def _set_input_dir(self, input, exclusive=True):
         """Set the input directory
