+++ jwst/stpipe/tests/test_format_template.py
